migrating files to the Swift package

+++ Sources/SpeculidKit/Extensions/CGFloat.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-extension CGFloat {
+extension Float {
   var cleanValue: String {
     truncatingRemainder(dividingBy: 1) == 0 ? String(format: "%.0f", self) : String(describing: self)
   }
